[IMP] scale: Provide functionality to generate notes from a scale.
This commit introduces the Scale object. The Scale object represents a
scale, initialized by its starting note and a vector of offsets. The
offsets indicate how the following notes of the scale are generated from
the base note. For example, C major can be represented as a base note
of C, and offsets of [0, 2, 4, 5, 7, 9, 11] corresponding to C, D, E, F,
G, A and B respectively.

Furthermore, this commit provides the ability to generate notes within a
scale, once the scale is initialized. The ScaleNote struct does this.

Internally, a ScaleNote merely contains the position of the generated
notes within the scale. When we want to render the note, a method is
provided to convert it into a NamedNote and a Note.

A NamedNote tracks not just the height of the note, but also how it
should be named (e.g. some scales use C#, others use Db).
This is mostly useful for outputting a human-readable representation.

A Note simply tracks the absolute height of a note, following the MIDI
convention.

By way of explanation, this commit introduces structs representing
'Notes' and 'Keys'. A Note is an absolute value on the keyboard;
a Key is a 'C', a 'D', etc. C3 and C4 are two distinct notes, but they
both resolve to the same key.