[IMP] json_input: Modification to input format duration construct
Previously, a note of specified duration would have been written
{"note": duration}.

We now invert this so that it is written {"duration": note}.

This has several advantages:
- an improved syntax for durations taking any "a/b" expression where
a is a numerator and b is a denominator, enabling e.g. triple-notes
to be written.
- the {"duration": note} syntax is recursive, allowing you to write
several notes at once (e.g. {"duration": [1, 2, 3]}). Note that in this
case the durations of [1, 2, 3] are not halved, unlike what would be
the case if you had just written [1, 2, 3] outside of a duration
construct.